+++ CHANGELOG.md
@@ -4,6 +4,66 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.3.1] - 2018-11-14
+### Changed
+- fixed issue of empty product URL keys
+- fixed double slashes issue
+- update category attributes via resource saveAttribute()
+- use proxy for CategoryUrlPathGenerator
+
+## [1.3.0] - 2018-10-29
+### Added
+- new option "--no-cache-clean"
+- new option "--no-cache-flush"
+- new option "--no-progress"
+- new option "--no-clean-url-key"
+
+### Changed
+- optimized code
+- modified logic of url rewrites regeneration
+- fixed issue of store filter in categories collection
+
+## [1.2.3] - 2018-10-03
+### Added
+- display additional debug information for "URL key for specified store already exists" error
+
+### Changed
+- modified logic of url rewrites regeneration
+
+## [1.2.2] - 2018-10-02
+### Changed
+- fixed setStoreId() on null error
+
+## [1.2.1] - 2018-09-25
+### Changed
+- fixed compilation issues
+
+## [1.2.0] - 2018-09-25
+### Changed
+- added proxies to CLI commands
+- modified logic of url rewrites regeneration
+- updated a composer file
+- fixed issue of a compatibility with new Magento Commerce versions
+
+## [1.1.1] - 2018-09-10
+### Changed
+- fix composer file format issue
+
+## [1.1.0] - 2018-09-09
+### Added
+- added feature to add a Pro features through a "Layer" class
+
+### Changed
+- fix issue when optional arguments require value
+- updated a code structure
+
+## [1.0.6] - 2018-07-26
+### Added
+- new option to run URL rewrite generation without running full reindex
+
+### Changed
+- update help notice to show INPUT_KEY_SAVE_REWRITES_HISTORY and INPUT_KEY_NO_REINDEX
+
 ## [1.0.5] - 2018-05-13
 ### Added
 - new option to save current URL rewrites